requirements.txt
django==4.1.1
djangorestframework==3.14.0
django-debug-toolbar==3.7.0
psycopg2>=2.8
